style🎨:Home 헤더 스타일링
