+++ src/pages/Homeview.jsx
@@ -3,6 +3,7 @@ import GenreChoice from 'components/Home/Content/GenreChoice';
 import RealtimeRank from 'components/Home/Content/RealtimeRank';
 import RecommendTicket from 'components/Home/Content/RecommendTicket';
 import TicketSlider from 'components/Home/Content/TicketSlider';
+import HomeviewHeader from 'components/Home/Header/HomeviewHeader';
 import useAPI from 'cores/hooks/useAPI';
 
 function Home() {
@@ -14,6 +15,7 @@ function Home() {
         <div>로딩중</div>
       ) : (
         <>
+          <HomeviewHeader />
           <TicketSlider bannerData={homeData.bannerList} />
           <GenreChoice />
           <RealtimeRank ticketData={homeData.rankList} />
